# Prototype distributed domains caching