Creating distributed domains has been optimized over the last several Chapel releases, but there is still a non-trivial creation overhead.

This prototypes caching distributed domains in a map to avoid recreating them each time. This has some minor performance benefits to SymEntry creation, which results in a ~5% speedup for things like stream and dataframe-display.

For now I just wanted to prototype this to evaluate performance, but I think if we want something like this it shouldn't be an unbounded cache, but some sort of LRU cache that only stores a couple of