@ronawho 2022-11-29
16-node-cs-hdr perf results for a couple benchmarks:

```
./benchmarks/run_benchmarks.py stream array_create dataframe -nl 16 --gen-graph --numtrials 3
```

Before:
```
==> benchmarks/datdir/array_create.dat <==
# Date	zeros Average time =	zeros Average rate =	ones Average time =	ones Average rate =	randint Average time =	randint Average rate =
11/29/22	0.0188	633.14	0.0210	568.11	0.0235	507.40
11/29/22	0.0136	875.08	0.0198	601.48	0.0250	477.58
11/29/22	0.0113	1051.68	0.0193	617.97	0.0200	597.32

==> benchmarks/datdir/dataframe.dat <==
# Date	_get_head_tail_server Average time =	_get_head_tail_server Average rate =	_get_head_tail Average time =	_get_head_tail Average rate =
11/29/22	0.4725	0.10	0.5208	0.09
11/29/22	0.3980	0.12	0.4646	0.10
11/29/22	0.3901	0.12	0.4357	0.11

==> benchmarks/datdir/stream.dat <==
# Date	Average time =	Average rate =
11/29/22	0.0729	490.63
11/29/22	0.0617	579.65
11/29/22	0.0560	639.19
```

After:
```
==> benchmarks/datdir/array_create.dat <==
# Date	zeros Average time =	zeros Average rate =	ones Average time =	ones Average rate =	randint Average time =	randint Average rate =
11/29/22	0.0129	923.21	0.0202	589.08	0.0233	512.49
11/29/22	0.0136	878.12	0.0202	589.30	0.0191	623.32
11/29/22	0.0147	810.81	0.0217	549.29	0.0191	625.01

==> benchmarks/datdir/dataframe.dat <==
# Date	_get_head_tail_server Average time =	_get_head_tail_server Average rate =	_get_head_tail Average time =	_get_head_tail Average rate =
11/29/22	0.3569	0.13	0.4184	0.11
11/29/22	0.3571	0.13	0.3481	0.13
11/29/22	0.3685	0.13	0.3755	0.12

==> benchmarks/datdir/stream.dat <==
# Date	Average time =	Average rate =
11/29/22	0.0624	573.20
11/29/22	0.0573	624.23
11/29/22	0.0547	654.04
```