@ronawho 2022-11-29
Just wanted to capture somewhere more permanent than a branch. Given the relatively small performance improvement, I don't plan to look at this in the near future but I think it would be nice to get to eventually.